src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    ./update_requirements_txt.sh
#
alembic==1.16.5
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anthropic==0.51.0
    # via -r requirements.in
anyio==4.11.0
    # via
    #   anthropic
    #   httpx
    #   mcp
    #   openai
    #   sse-starlette
    #   starlette
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via posthog
boto3==1.40.50
    # via -r requirements.in
botocore==1.40.50
    # via
    #   boto3
    #   s3transfer
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   typer
    #   uvicorn
distro==1.9.0
    # via
    #   anthropic
    #   openai
    #   posthog
fastapi==0.118.3
    # via
    #   -r requirements.in
    #   fastapi-pagination
fastapi-pagination==0.14.3
    # via -r requirements.in
greenlet==3.2.4
    # via sqlalchemy
grpcio==1.75.1
    # via qdrant-client
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.28.1
    # via
    #   -r requirements.in
    #   anthropic
    #   mcp
    #   ollama
    #   openai
    #   qdrant-client
httpx-sse==0.4.3
    # via mcp
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.11.0
    # via
    #   anthropic
    #   openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.25.1
    # via mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via mako
mcp[cli]==1.17.0
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
mem0ai==0.1.118
    # via -r requirements.in
numpy==2.3.3
    # via qdrant-client
ollama==0.4.8
    # via -r requirements.in
openai==1.109.1
    # via
    #   -r requirements.in
    #   mem0ai
portalocker==3.2.0
    # via qdrant-client
posthog==6.7.6
    # via mem0ai
protobuf==5.29.5
    # via
    #   mem0ai
    #   qdrant-client
psycopg2-binary==2.9.11
    # via -r requirements.in
pydantic==2.12.0
    # via
    #   anthropic
    #   fastapi
    #   fastapi-pagination
    #   mcp
    #   mem0ai
    #   ollama
    #   openai
    #   pydantic-settings
    #   qdrant-client
pydantic-core==2.41.1
    # via pydantic
pydantic-settings==2.11.0
    # via mcp
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   posthog
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   mcp
    #   pydantic-settings
python-json-logger==4.0.0
    # via -r requirements.in
python-multipart==0.0.20
    # via
    #   -r requirements.in
    #   mcp
pytz==2025.2
    # via mem0ai
qdrant-client==1.15.1
    # via mem0ai
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via posthog
rich==14.2.0
    # via typer
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
s3transfer==0.14.0
    # via boto3
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   posthog
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   openai
sqlalchemy==2.0.44
    # via
    #   -r requirements.in
    #   alembic
    #   mem0ai
sse-starlette==3.0.2
    # via mcp
starlette==0.48.0
    # via
    #   fastapi
    #   mcp
tenacity==9.1.2
    # via -r requirements.in
tqdm==4.67.1
    # via openai
typer==0.19.2
    # via mcp
typing-extensions==4.15.0
    # via
    #   alembic
    #   anthropic
    #   anyio
    #   fastapi
    #   fastapi-pagination
    #   grpcio
    #   openai
    #   posthog
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via
    #   botocore
    #   qdrant-client
    #   requests
uvicorn==0.37.0
    # via
    #   -r requirements.in
    #   mcp
apscheduler>=3.10.0
